Name of folder and components for HowClient.page have been changed to HowClientDescription. Styling adjusted to move content on the right place on the page
