+++ src/components/HowClientDescription/HowClientDescription.styled.ts
@@ -1,12 +1,11 @@
 import styled from "styled-components";
 
 export const SearchPageDescription = styled.div`
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%);
-
   display: flex;
-  justify-content: center;
+  flex-direction: column;
+  /* justify-content: center; */
   align-items: center;
   text-align: center;
+
+  width: 557px;
 `;
