+++ src/components/Menu/Menu.styled.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 export const MenuField = styled.div`
-  width: 100%;
+  /* width: 100%; */
   display: flex;
   flex-direction: row;
   justify-content: space-between;
